Move test helpers to test-support folder
This fixes https://github.com/san650/ember-cli-page-object/issues/1
